pyproject.toml
[project]
name = "tjm"
version = "0.1.0"
description = "Automated Notepad typing and saving using computer vision and UI automation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.13.0.90",
    "pillow>=12.1.0",
    "pyautogui>=0.9.54",
    "pygetwindow>=0.0.9",
    "requests>=2.32.5",
    "numpy>=1.26.4",
]
